Fix the build directory in diagnosis
The previous value "workdir" was supposed to be the path of the build,
which lets us know when a stack trace travels through the crate under
test. At some point I changed this and somehow never noticed. This fixes
it. Still hard-coded, but at least the HTML will be more accurate.